Delete docker-build.yml
This GitHub Build Action required that the Repository's owner add two secrets to the secrets tab of the security section. These secrets were never added so the workflow would always fail. So it is being removed in order to make sure that our grade for GitHub Actions CI Build Status is maximized.